feat: add exception and filter

+++ src/main/java/vn/ptit/controller/manufacturer/DeleteManufactureController.java
@@ -4,6 +4,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import vn.ptit.controller.ResponseBody;
+import vn.ptit.exception.InvalidRequestException;
 import vn.ptit.model.PagingPayload;
 import vn.ptit.service.manufacturer.DeleteManufacturerService;
 import vn.ptit.service.manufacturer.UpdateManufacturerService;
@@ -19,6 +20,9 @@ public class DeleteManufactureController {
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<?> update(@PathVariable("id") Long id){
+        if (id == null)
+            throw new InvalidRequestException("Require [id]");
+
         deleteManufacturerService.delete(id);
         return new ResponseEntity<>(new ResponseBody(PagingPayload.empty(),ResponseBody.Status.SUCCESS,ResponseBody.Code.SUCCESS), HttpStatus.OK);
     }
