+++ src/main/java/vn/ptit/service/manufacturer/DeleteManufacturerService.java
@@ -1,6 +1,8 @@
 package vn.ptit.service.manufacturer;
 
+import lombok.SneakyThrows;
 import org.springframework.stereotype.Service;
+import vn.ptit.exception.DataNotFoundException;
 import vn.ptit.model.Manufacturer;
 import vn.ptit.repository.manufacturer.IManufacturerRepository;
 
@@ -11,9 +13,12 @@ public class DeleteManufacturerService {
     public DeleteManufacturerService(IManufacturerRepository manufacturerRepository) {
         this.manufacturerRepository = manufacturerRepository;
     }
-
+    @SneakyThrows
     public void delete(Long id){
         Manufacturer manufacturer = manufacturerRepository.getById(id);
+        if (manufacturer == null) {
+            throw new DataNotFoundException("Manufacturer not found");
+        }
         manufacturer.delete();
         manufacturerRepository.save(manufacturer);
     }
